moved all the project view material together
