--- client/src/components/searchcard/searchcard.css
@@ -1,10 +1,0 @@
-.search {
-  display: flex;
-  justify-content: center;
-  padding: 20px 0;
-}
-
-.card {
-  display: flex;
-  justify-content: center;
-}
