--- client/src/pages/Search.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SearchContainer from "../components/Searchcontainer/searchcontainer";
-
-function Search() {
-	return (
-		<div className="container">
-			<SearchContainer />
-		</div>
-	);
-};
-
-export default Search;
